a4/requirements.txt
matplotlib.pyplot
networkx
sys
numpy